Port Lesson 8 Game over (not exactly)
Phaser 3 doesn't seem to have an API to detect a sprite gets out of bounds of the world (Ref: https://phaser.discourse.group/t/out-of-bounds-checking/905). So I decided to change the logic of the game so that the game is over when the ball touches the bottom so far.

This would be sufficient for this game, but I'll try another way later to decide which to adopt